IEEE 754 Single-Precision floating-point format "float" can represent more integer numerical values than "int", but it reserves 23 bits in mantissa field. Whereas there are 32 bits in "int". Hence, "float" only guarantee 1-to-1 mapping up to (2 ^ 24 - 1) with the implicit one in definition of mantissa field, and thus when type casting from "int" to "float" may encounter some precision problem.
William Kahan demostrated this problem and provided draft solution to
fix the precision. I think that the bias from IEEE 754 precsion can
come from two possible sources, i.e., the arithmetic during calculation,
and the accommodation status after calculation. For the first source,
arithmetic during calculation occurs with rounding the mantissa field
when two "float" variables are aligned with the same exponent field.
For the other source, the 23 bits in mantissa field cannot precisely
represent "int" variable with too large values and causes the rounding.

I am interested in the exact bias in the demo of Kahan Summation. I try
to evaluate the bias, but only limited to the source come from the
aforementioned second source of rounding. This is, this may not work in
other scenarios.