Define the methodology
How to estimate the precision error in the scenario of accumulating from
1 to 10000 to "float"? It yields 50002896.000000 rather than the correct
answer 50005000.

Feasibility: Within accumulation results of these 10000 rounds, the
precision error may appear as the accumulation result is greater than
2 ^ 24, which occupies 25 bits to present its value without precision
error and may not precisely fit into "float" variable. The most
significant bits will be kept, and remaining bits will be rounded.
Besides, "int" variable has 31 bits to represent the positive integer
and zero. If we understand how the rounding method of "float" operates,
it is possible to use "int" estimating the precision errors.

Challenge: Once the precision error appear, it will propagate into the
further rounds. Thus, in my implementation, I inject the precision error
into each round, and collect the final precision error as the estimation.

   accumulation --- error injection ---------> current precision error
       ^                               |
       |_______________________________|
                   next round?